samples: Add NFC TNEP Connection Handover to the NFC Pairing sample
The NFC Pairing sample is renamed to the Peripheral NFC
Pairing.

NCSDK-5094
