+++ samples/bluetooth/peripheral_nfc_pairing/CMakeLists.txt
