+++ samples/bluetooth/peripheral_nfc_pairing/Kconfig
@@ -1,0 +1,37 @@
+#
+# Copyright (c) 2020 Nordic Semiconductor
+#
+# SPDX-License-Identifier: LicenseRef-BSD-5-Clause-Nordic
+#
+
+menu "NFC Paring configuration"
+
+choice NFC_TAG_CONNECTION_HANDOVER_ROLE
+	prompt "Connection Handover Tag role"
+	default NFC_TAG_CH_SELECTOR
+
+config NFC_TAG_CH_SELECTOR
+	bool "Connection Handover selector role"
+	help
+	  NFC Tag is Connection Handover selector device. The Handover Selector
+	  is an NFC Forum Device that constructs and replies to a
+	  Handover Select Message as a result of a previously received
+	  Handover Request Message, or an NFC Tag Device that provides a
+	  pre-set Handover Select Message for reading.For more details
+	  check NFC Forum Connection Handover 1.5 2.2
+
+config NFC_TAG_CH_REQUESTER
+	bool "Connection Handover requester role"
+	help
+	  NFC Tag is Connection Handover requester device. The Handover Requester
+	  is an NFC Forum Device that begins the Handover Protocol by issuing
+	  a Handover Request Message to another NFC Forum Device. For more
+	  details check NFC Forum Connection Handover 1.5 2.2.
+
+endchoice
+
+endmenu
+
+menu "Zephyr Kernel"
+source "Kconfig.zephyr"
+endmenu
