+++ samples/bluetooth/peripheral_nfc_pairing/prj.conf
@@ -15,11 +15,19 @@ CONFIG_NFC_NDEF_RECORD=y
 CONFIG_NFC_NDEF_LE_OOB_REC=y
 CONFIG_NFC_NDEF_CH_MSG=y
 
+CONFIG_NFC_NDEF_TNEP_RECORD=y
+CONFIG_NFC_TNEP_TAG=y
+CONFIG_NFC_NDEF_PARSER=y
+CONFIG_NFC_NDEF_CH_PARSER=y
+CONFIG_NFC_NDEF_LE_OOB_REC_PARSER=y
+CONFIG_NFC_TNEP_CH=y
+
 CONFIG_BT=y
 CONFIG_BT_SMP=y
 CONFIG_BT_SMP_APP_PAIRING_ACCEPT=y
 CONFIG_BT_PERIPHERAL=y
 CONFIG_BT_DEVICE_NAME="Nordic_NFC_pairing"
+CONFIG_BT_TINYCRYPT_ECC=y
 
 CONFIG_BT_DIS=y
 CONFIG_BT_DIS_PNP=y
@@ -30,6 +38,7 @@ CONFIG_BT_DIS_PNP_PID=0xEEEF
 CONFIG_BT_DIS_PNP_VER=0x0100
 
 CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=2048
+CONFIG_MAIN_STACK_SIZE=2048
 
 CONFIG_BT_SETTINGS=y
 CONFIG_FLASH=y
@@ -39,3 +48,5 @@ CONFIG_NVS=y
 CONFIG_SETTINGS=y
 
 CONFIG_DK_LIBRARY=y
+
+CONFIG_POLL=y
