+++ samples/bluetooth/peripheral_nfc_pairing/sample.yaml
@@ -1,0 +1,9 @@
+sample:
+  name: Bluetooth Low Energy Peripheral NFC Pairing reference
+  description: Bluetooth Low Energy Peripheral NFC Pairing reference sample
+tests:
+  samples.bluetooth.peripheral_nfc_pairing:
+    build_only: true
+    build_on_all: true
+    platform_allow: nrf52840_pca10056 nrf52_pca10040 nrf5340_dk_nrf5340_cpuapp
+    tags: bluetooth ci_build
